% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{bias_plot}
\alias{bias_plot}
\title{Plot Bias Histogram}
\usage{
bias_plot(bias = NULL, save_plot = FALSE, file_path = NULL, w = 6.5, h = 6.5)
}
\arguments{
\item{bias}{Numeric vector containing values of relative bias between posterior means and true value of the % population decline per 10 years across all simulated data sets.}

\item{save_plot}{Logical value indicating whether to save plot.}

\item{file_path}{File path including user-provided file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}
}
\value{

}
\description{
Plot Bias Histogram
}
