% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ci_cover_plot}
\alias{ci_cover_plot}
\title{Plot Bayesian Credible Intervals for All Simulations}
\usage{
ci_cover_plot(
  pctdecl_ests = NULL,
  cover_95 = NULL,
  true_pctdecl = 25,
  save_plot = FALSE,
  file_path = NULL,
  w = 6.5,
  h = 9
)
}
\arguments{
\item{pctdecl_ests}{Matrix containing summary statistics calculated from posterior distributions of the percent population decline per 10 years for all simulated data sets.}

\item{cover_95}{Numeric vector containing posterior distributions means of the percent population decline per 10 years for all simulated data sets.}

\item{true_pctdecl}{Numeric value specifying true percent population decline per 10 years used for simulations.}

\item{save_plot}{Logical value indicating whether to save plot.}

\item{file_path}{File path including user-provided file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}
}
\value{

}
\description{
Plot Bayesian Credible Intervals for All Simulations
}
