% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{create_folders}
\alias{create_folders}
\title{Create Folders}
\usage{
create_folders(dir = NULL, moreFolders = NULL)
}
\arguments{
\item{dir}{Location to create new 'data', 'models', and 'output' folders.}

\item{moreFolders}{Character vector of additional folder names to be created.}
}
\description{
Create Folders
}
