% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cv_plot}
\alias{cv_plot}
\title{Plot Coefficient of Variation Histogram}
\usage{
cv_plot(cv = NULL, save_plot = FALSE, file_path = NULL, w = 6.5, h = 6.5)
}
\arguments{
\item{cv}{Numeric vector containing coefficient of variation (CV) estimates calculated from posterior distributions for the % population decline per 10 years across all simulated data sets.}

\item{save_plot}{Logical value indicating whether to save plot.}

\item{file_path}{File path including user-provided file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}
}
\value{

}
\description{
Plot Coefficient of Variation Histogram
}
