% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{density_diff_plot}
\alias{density_diff_plot}
\title{Plot Posterior of Density Difference Between Islands}
\usage{
density_diff_plot(
  mod_output = NULL,
  save_plot = FALSE,
  file_path = NULL,
  w = 6.5,
  h = 6.5
)
}
\arguments{
\item{mod_output}{MCMC output object of class \code{jagsUI} from either model.}

\item{save_plot}{Logical value indicating whether to save plot.}

\item{file_path}{File path including user-provided file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}
}
\value{

}
\description{
Plot Posterior of Density Difference Between Islands
}
