% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{detfn_plot}
\alias{detfn_plot}
\title{Plot Distance Detection Function}
\usage{
detfn_plot(
  mod3_output = NULL,
  mod4_output = NULL,
  save_plot = FALSE,
  file_path = NULL,
  w = 9,
  h = 6.5
)
}
\arguments{
\item{mod3_output}{MCMC output object of class \code{jagsUI} for model with groups as observation unit.}

\item{mod4_output}{MCMC output object of class \code{jagsUI} for model with individuals as observation unit.}

\item{save_plot}{Logical value indicating whether to save plot.}

\item{file_path}{File path including user-provided file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}
}
\value{

}
\description{
Function plots distance detection functions for individual models (i.e., input provided for 1 model argument)
or functions from both models (i.e., input provided for both model argument) overlaid on the sample plot.
}
