% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{distdetfn}
\alias{distdetfn}
\title{Calculate Distance Detection Probabilities}
\usage{
distdetfn(dists = NULL, sigma = NULL)
}
\arguments{
\item{dists}{Numeric vector of distances from transect line.}

\item{sigma}{Spatial-scale parameter for Gaussian (half-normal) distance detection function.}
}
\value{
Numeric vector of detection probabilities.
}
\description{
Calculate Distance Detection Probabilities
}
