% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{EN_isl_plot}
\alias{EN_isl_plot}
\title{Plot Posterior Sample Histograms for Expected Abundance}
\usage{
EN_isl_plot(
  mod_output = NULL,
  estimate = "total",
  year = NULL,
  save_plot = FALSE,
  file_path = NULL,
  w = 9,
  h = 6.5
)
}
\arguments{
\item{mod_output}{MCMC output object of class \code{jagsUI} from either model.}

\item{estimate}{Character vector specifying one or more island estimates to plot. Accepted values are "total", "hall", and "stma" and default is "total".}

\item{year}{Numeric vector specifying years for which island estimates are plotted. Accepted values are  2003, 2018, and 2028 and the default is all 3.}

\item{file_path}{File path including user-provided file name. Observer indexes will automatically be appended to file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}

\item{save_plots}{Logical value indicating whether to save plots.}
}
\value{

}
\description{
Plot Posterior Sample Histograms for Expected Abundance
}
