% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_3.R
\name{fit_model3}
\alias{fit_model3}
\title{Fit JAGS Model 3}
\usage{
fit_model3(
  mcbu_data_file = "data/distdata_mcbu.csv",
  ni = 200,
  nb = 100,
  nt = 1,
  nc = 1,
  M = c(200, 100),
  save_output = FALSE,
  jags_output_folder = NULL,
  use_parallel = TRUE,
  ...
)
}
\arguments{
\item{mcbu_data_file}{File path and name for MCBU distance data. Default is set to look for "distdata_mcbu.csv" in the "data" folder created by the \code{create_folders} function.}

\item{ni}{Total number of MCMC samples to simulate including burn-in samples. (Does not include adaptive samples)}

\item{nb}{Number of MCMC samples to discard as burn-in samples.}

\item{nt}{Integer specifying the thinning rate for MCMC samples.}

\item{nc}{Integer specifying the number of individual MCMC chains to simulate.}

\item{M}{Integer vector specifying the augmented number of MCBU groups per square kilometer.}

\item{save_output}{Logical value specifying whether to save JAGS output.}

\item{jags_output_folder}{Folder location to save JAGS output. Default is the "output" folder created by the \code{create_folders} function.}

\item{use_parallel}{Logical value specifying whether to use parallel computing when simulating >1 MCMC chain. See help for \code{jags} function in \code{jagsUI} package for more details.}
}
\value{
JAGS MCMC output from fitting an exponential growth model with individual groups as observation unit and observer random effect on sigma parameter to MCBU distance-sampling data from 2003 and 2018. Output is saved as an RData file named \code{out3.RData}.
}
\description{
Fit JAGS Model 3
}
