% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{getEN_mcmc}
\alias{getEN_mcmc}
\title{Extract MCMC Samples for Expected N}
\usage{
getEN_mcmc(mod_output = NULL)
}
\arguments{
\item{mod_output}{MCMC output object of class \code{jagsUI} from either model.}
}
\value{
Array containing MCMC samples for island-specific and total abundance for surveyed years (2003 and 2018) and projected abundances for 2028.
}
\description{
Extract MCMC Samples for Expected N
}
