% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{growthN_plot}
\alias{growthN_plot}
\title{Plot Population Growth From 2003 to 2018}
\usage{
growthN_plot(
  mod_output = NULL,
  estimate = c("pctdecline", "r", "lambda"),
  save_plot = FALSE,
  file_path = NULL,
  w = 6.5,
  h = 9
)
}
\arguments{
\item{mod_output}{MCMC output object of class \code{jagsUI} from either model.}

\item{estimate}{Character vector specifying one or more estimates of population growth to plot. Accepted values are "pctdecline" for percent population decline over the study period (15 years), "r" for annual exponential population growth rate, and "lambda" for annual geometric population growth rate. Default returns plots for all 3.}

\item{file_path}{File path to save figures. File names are provided.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}

\item{save_plots}{Logical value indicating whether to save plots.}
}
\value{

}
\description{
Plot Population Growth From 2003 to 2018
}
