% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{Mode}
\alias{Mode}
\title{Calculate Mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
numeric value of mode as estimated by the base \code{density} function in R with \code{na.rm = TRUE}
}
\description{
Calculate Mode
}
