% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mcbu_power.R
\name{process_mcbu_power}
\alias{process_mcbu_power}
\title{Process and Summarize MCBU Power Simulation Output}
\usage{
process_mcbu_power(
  true_pctdecl = 25,
  output_filepath = NULL,
  save_figures = TRUE,
  figure_filepath = NULL,
  save_summary = TRUE,
  summary_filepath = NULL
)
}
\arguments{
\item{true_pctdecl}{True percent decline over 10 years used in simulations.}

\item{output_filepath}{File path to folder containing saved JAGS output from power simulations.}

\item{save_figures}{Logical value specifying whether to save summary figures.}

\item{figure_filepath}{File path to folder to which figures are saved.}

\item{save_summary}{Logical value specifying whether to save summary R objects.}

\item{summary_filepath}{File path to folder to which summary R objects are save.}
}
\value{
None
}
\description{
Process and Summarize MCBU Power Simulation Output
}
