% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcbu_power.R
\name{run_mcbu_power}
\alias{run_mcbu_power}
\title{MCBU Model 4 Power Analysis - Data and MCMC Simulator}
\usage{
run_mcbu_power(
  log_lambda0 = c(4.1257, 3.94),
  r = log(0.75^(1/10)),
  mu_obs = -2.9769,
  sd_obs = 0.447,
  survey_years = c(2018, 2028),
  mcbu_data_file = "distdata_mcbu.csv",
  save_output = FALSE,
  sims_output_folder = NULL,
  batch = 1,
  M = c(200, 100),
  ni = 200,
  nb = 100,
  nt = 1,
  nc = 1,
  use_parallel = TRUE,
  ...
)
}
\arguments{
\item{log_lambda0}{Numeric vector of length 2 containing starting island-specific (\code{c(Hall,St Matthew)}) MCBU densities in square km. Default set to preliminary estimates for 2018.}

\item{r}{Numeric value specifying exponential growth rate \code{r}. Default value is set to 25 percent decline over 10 years.}

\item{mu_obs}{Numeric value specifying the mean sigma value across observers on the log scale. Default set to preliminary estimates for 2003 and 2018.}

\item{sd_obs}{Numeric value specifying the standard deviation of sigma values across observers on the log scale. Default set to preliminary estimates for 2003 and 2018.}

\item{survey_years}{Numeric vector specifying in which years surveys will be conducted.}

\item{mcbu_data_file}{File path and name for MCBU distance data. Default is set to distdata_mcbu.csv in the working directory.}

\item{save_output}{Logical value specifying whether to save JAGS output. Default set to FALSE.}

\item{sims_output_folder}{Folder location to save JAGS output from simulations. Default creates an "output/'design'/mcmc" subdirectory in working directory.}

\item{batch}{Integer vector specifying replicates for data and MCMC simulation. Default set to single replicate.}

\item{M}{Integer vector specifying the augmented number of MCBU groups per square kilometer.}

\item{nb}{Number of MCMC samples to discard as burn-in samples.}

\item{nt}{Integer specifying the thinning rate for MCMC samples.}

\item{nc}{Integer specifying the number of individual MCMC chains to simulate.}

\item{use_parallel}{Logical value specifying whether to use parallel computing when simulating >1 MCMC chain. See help for \code{jags} function in \code{jagsUI} package for more details.}
}
\value{
No output objects are returned.
}
\description{
MCBU Model 4 Power Analysis - Data and MCMC Simulator
}
