% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sq_err_plot}
\alias{sq_err_plot}
\title{Plot Squared Error Histogram}
\usage{
sq_err_plot(
  sq_err = NULL,
  save_plot = FALSE,
  file_path = NULL,
  w = 6.5,
  h = 6.5
)
}
\arguments{
\item{sq_err}{Numeric vector containing squared error estimates calculated from posterior distributions for the percent population decline per 10 years across all simulated data sets.}

\item{save_plot}{Logical value indicating whether to save plot.}

\item{file_path}{File path including user-provided file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}
}
\value{

}
\description{
Plot Squared Error Histogram
}
