% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{summary.fn}
\alias{summary.fn}
\title{Calculate Simple Summary Statistics}
\usage{
\method{summary}{fn}(x = NULL)
}
\arguments{
\item{x}{numeric vector or matrix}
}
\value{
numeric vector or matrix when x is supplied. Returns labels when x is NULL.
}
\description{
Calculate Simple Summary Statistics
}
