% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_1.R
\name{write_model_1}
\alias{write_model_1}
\title{Write JAGS Model 1 to File}
\usage{
write_model_1(model_filepath = "./models/model_1.txt")
}
\arguments{
\item{model_filepath}{File path and name for JAGS model to be written}
}
\value{
Text file containing JAGS model is saved to \code{model_filepath}
}
\description{
Write JAGS Model 1 to File
}
