% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2.R
\name{write_model_2}
\alias{write_model_2}
\title{Write JAGS Model 2 to File}
\usage{
write_model_2(model_filepath = NULL)
}
\arguments{
\item{model_filepath}{File path and name for JAGS model to be written}
}
\value{
Text file containing JAGS model is saved to \code{model_filepath}
}
\description{
Write JAGS Model 2 to File
}
