% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_4.R
\name{write_model_4}
\alias{write_model_4}
\title{Write JAGS Model 4 to File}
\usage{
write_model_4(model_filepath = NULL)
}
\arguments{
\item{model_filepath}{File path and name for JAGS model to be written. Default is the "models" folder created by the \code{create_folders} function.}
}
\value{
Text file containing JAGS model is saved to \code{model_filepath}
}
\description{
Write JAGS Model 4 to File
}
