% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{obs_detfn_plot}
\alias{obs_detfn_plot}
\title{Plot Observer-Specific Detection Functions and Observations}
\usage{
obs_detfn_plot(
  mcbu_data_file = "data/distdata_mcbu.csv",
  mod_output = NULL,
  save_plots = FALSE,
  file_path = NULL,
  w = 6.5,
  h = 6.5
)
}
\arguments{
\item{mcbu_data_file}{File path and name for MCBU distance data. Default is set to look for "distdata_mcbu.csv" in the "data" folder created by the \code{create_folders} function.}

\item{mod_output}{MCMC output object of class \code{jagsUI} from either model.}

\item{save_plots}{Logical value indicating whether to save plots.}

\item{file_path}{File path including user-provided file name. Observer indexes will automatically be appended to file name.}

\item{w}{Numeric value specifying figure width in inches.}

\item{h}{Numeric value specifying figure height in inches.}
}
\value{

}
\description{
Plot Observer-Specific Detection Functions and Observations
}
